documentation_frontend/documentation_frontend/requirements.txt
djangorestframework==3.14.0
django-cors-headers
psycopg2-binary
python-decouple